refactor hf_xet library into data crate
