+++ data/Cargo.toml
@@ -37,6 +37,8 @@ toml = "0.5"
 static_assertions = "1.1.0"
 gearhash = "0.1.3"
 rand_chacha = "0.3.1"
+dirs = "5.0.1"
+async-once-cell = "0.5.4"
 
 # metrics
 prometheus = "0.13.0"
